feat: calculate latency

+++ trading/trading-types/src/from_server.rs
@@ -21,7 +21,7 @@ pub struct TickData {
 
 #[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub struct Latency {
-    pub ms: u32,
+    pub ms: u64,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
